IoT_Gateway/requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
websockets==11.0.3